Update Recommendation 3 to avoid ambiguity of roles
In recommendation 3, the roles of Counselors and Ombudspersons have been blurred. We already have established roles for the ombudsperson who are external to the community, and recommendation 3 should highlight that we should have counsellors from within the community to facilitate CoC work. 

These points have been raised in the CoC facilitators draft by me and Karin Lagesen and suggested by @serahrono I am opening this PR for @kariljordan's review.